Change distance calculation to degrees
